refactor Refac continues. Moved JPG video script to Object.
